Rename ValidatorSimple to Validator
