--- src/renderer/components/Validator/validatorsimple.scss
@@ -1,47 +1,0 @@
-.validator-simple-container{
-    background-color: $neutral-700;
-    box-shadow: 0px 2px 56px rgba(0, 0, 0, 0.3);
-    border-radius: 24px;
-    padding: 32px;
-    max-width: 1280px;
-    display: flex;
-    flex: 1;
-    justify-content: space-between;
-}
-    .validator-simple-keys{
-        // width: 50%;
-        width: 800px;
-        display: flex;
-        flex-direction: column;
-        justify-content: space-between;
-
-        .validator-stat-container {
-            justify-content: space-evenly;
-
-            .validator-card-container:nth-child(2) {
-                margin: 75px 0 0;
-            }
-        }
-    }
-    .public-key{
-        padding: 32px 0 0 10px;
-    }
-
-    .validator-simple-status{
-        display: flex;
-        flex-direction: column;
-        align-items: flex-end;
-        justify-content: space-between;
-    }
-        .validator-simple-status h2{
-            margin-bottom: 10px;
-            align-self: baseline;
-        }
-        .validator-simple-buttons{
-            margin-top: 25px;
-            display: flex;
-            justify-content: flex-end;
-            .button:nth-child(1){
-                margin-right: 27px;
-            }
-        }
