+++ src/renderer/style/index.scss
@@ -16,7 +16,6 @@
 @import '../components/Notification/notification.scss';
 @import '../components/VerifyMnemonic/VerifyMnemonic.scss';
 @import '../components/Validator/validator.scss';
-@import '../components/Validator/validatorsimple.scss';
 @import '../components/Cards/cards.scss';
 @import '../components/Loading/loading.scss';
 @import '../components/LineGraph/linegraph.scss';
